size_t is not the same as unsigned on 64bits
git-svn-id: file:///home/luispedro/luispedro.org/svn/hearts/trunk@474 174c993c-7530-0410-8f2a-93c0592b7a1f

+++ hearts/common/network/message.h
@@ -191,7 +191,7 @@ inline const Message::errorType to<Message::errorType>( const QString& r ) {
 }
 
 template <typename T>
-inline const T Message::arg( size_t idx ) const
+inline const T Message::arg( unsigned idx ) const
 {
 	QString rep = arg( idx );
 	return to<T>( rep );
